// frontend/src/apisv2beta1/run/models/PipelineTaskTaskType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *  - TASK_TYPE_UNSPECIFIED: Default value. This value is not used.
 *  - ROOT: Root task is the top ancestor task to all tasks in the pipeline run
 * It is the only task with no parent task in a Pipeline Run.
 *  - RUNTIME: All child tasks in the Run DAG are Runtime tasks. With the exception
 * of K8S driver pods.
 * These tasks are the only tasks that have Executor Pods.
 *  - CONDITION_BRANCH: Condition Branch is the wrapper task of an If block
 *  - CONDITION: Condition is an individual "if" branch, and is
 * a child to a CONDITION_BRANCH task.
 *  - LOOP: Task Group for CONDITION_BRANCH
 * Task Group for RUNTIME Loop Iterations
 *  - DAG: Generic DAG task type for types like Nested Pipelines
 * where there is no declarative way to detect this within
 * a driver.
 * @export
 */
export const PipelineTaskTaskType = {
  TASK_TYPE_UNSPECIFIED: 'TASK_TYPE_UNSPECIFIED',
  ROOT: 'ROOT',
  RUNTIME: 'RUNTIME',
  CONDITION_BRANCH: 'CONDITION_BRANCH',
  CONDITION: 'CONDITION',
  LOOP: 'LOOP',
  EXIT_HANDLER: 'EXIT_HANDLER',
  IMPORTER: 'IMPORTER',
  DAG: 'DAG',
} as const;
export type PipelineTaskTaskType = (typeof PipelineTaskTaskType)[keyof typeof PipelineTaskTaskType];

export function instanceOfPipelineTaskTaskType(value: any): boolean {
  for (const key in PipelineTaskTaskType) {
    if (Object.prototype.hasOwnProperty.call(PipelineTaskTaskType, key)) {
      if (PipelineTaskTaskType[key as keyof typeof PipelineTaskTaskType] === value) {
        return true;
      }
    }
  }
  return false;
}

export function PipelineTaskTaskTypeFromJSON(json: any): PipelineTaskTaskType {
  return PipelineTaskTaskTypeFromJSONTyped(json, false);
}

export function PipelineTaskTaskTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskType {
  return json as PipelineTaskTaskType;
}

export function PipelineTaskTaskTypeToJSON(value?: PipelineTaskTaskType | null): any {
  return value as any;
}

export function PipelineTaskTaskTypeToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskType {
  return value as PipelineTaskTaskType;
}